pyproject.toml
[tool.poetry]
name = "numpy-cli"
version = "0.1.0"
description = ""
authors = ["Ben Skubi <skubi@ohsu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
numpy = "^1.26.3"
click = "^8.1.7"

[tool.poetry.scripts]
numpy-cli = "numpy_cli.main:cli"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
